requirements.txt
flask
tensorflow
Pillow
